tweaked values
tweaked max and min values, and inscale weights vs outscale weight, I think the rhythm is a BIT to similar, but the notes seem pretty good right now. Need to consider how first note is selected, cause right now if c's the first note it'll just take the one after c and that's a bit dull?